Flexible id list format.
The csv module is used to auto-detect the format of the file. Student ID
must be in the first column of the file. The rest of columns are
ignored. Students ID may be quoted or not. Any separator field (tab,
comma,...) is valid as long as the csv module can guess it.