Option to keep exam order for already shuffled exams.
Option -k in create_exam.py generates exams in the same order they are
in the exam config file, when the model to be created is already in that
file. If the option is not set, or the model is not in the exam config
file, the model is shuffled again.