Allowing setting IDs even when ID boxes must not be read.
In some circumstances there is no ID box but users should be able to
manually set a student ID nevertheless. The same happens when the
student ID box prevents the answer box from being read because of an
inappropriate geometry. In those cases eyegrade can be configured to not
read the ID, but let the user put manually.