First attempt at multiprocessing.
This code uses PyQT threading support. However, the interface is not
very responsive despite that. Maybe I should explore a solution that
uses another process instead of a thread.

In addition, there may be a problem if snapshot/manual detection are
selected during the processing of a capture. It might be safer in those
case to wait for the finishing of the thread.