Automatic detection of exam removal.
First attempt at issue #34. It is experimental by now. It can be tested
by activating the option "Continue on exam removal" in the "Tools" menu.

Testing is needed. One of the things to be tested is that the new exam
is not locked while being placed, before it is at its final position. In
addition, the Hough threshold should probably be locked until the new
exam is placed and stable.