Workaround for a problem with QThreadPool.
See issue #66: A problem with QThreadPool prevents eyegrade from
running when the session is open. The blue screen is shown in the
interface instead of the output of the camera. The problem happens
with PyQt version 4.10.1-1. If it is replaced by version 4.9.3-4,
eyegrade works fine. It may be a bug in PyQt. I'll investigate the
problem.

This commit is a temporary solution that tests QThreadPool and runs
the tasks inline instead of using QThreadPool when QThreadPool does
not work.