Grouping questions not to be shuffled by eyegrade-create
Questions in the input XML question file can now be grouped
with the "group" element. Groups get shuffled with the rest
of groups and standalone questions, but the order of questions
within the same group gets preserved.

This allows chaining related questions that form a logical sequence
in which order should be preserved and no other question
should be inserted in the middle.

This commit solves issue #145, but further testing and documentation
are needed before closing the issue.