PyQt5 5.15 can now be used
The 5.14 version is no longer pinned. There were two causes
for the problem:

1) Running it in an oldstable or oldoldstable Debian system
with libxcb-utils0 instead of libxcb-utils1.

2) That opencv-python ships a XCB QPA plugin that is incompatible
with PyQt5, and it is loaded instead of the version of the plugin
that comes with PyQt5: https://forum.qt.io/topic/119109/using-pyqt5-with-opencv-python-cv2-causes-error-could-not-load-qt-platform-plugin-xcb-even-though-it-was-found/2

Running a more recent system (the current Debian stable: bullseye)
and depending on opencv-python-headless instead of opencv-python
solves the problems.