Prevent actions until an exam is fully stored and added to the GUI
Otherwise, it's possible to remove an exam whose capture has not yet been
saved to disk and displayed in the interface, which causes trouble.

A new temporary mode mode called `before_review_from_grading`
has been added for the period before the capture has been saved
and the interface updated with it. After that, the `review_from_grading`
mode is entered and the user can proceed to interact with the exam.

Closes issues #161 and #127.