Add model variations to exam configuration
Variations have been split into question variations and model variations.
Question variations are the old variations, and they may vary question
by question. A model variation is a variation that applies to a whole model.
It affects the selection of question variations, but can also be used
to control variations in external questions that are mixed in the same
exam as the MCQ questions. The {{variation}} LaTeX parameter
stores now the model variation.

After this change, it is possible to use model variations even if no
MCQ question in the exam defines variations.