Change how students are shown in student id dialog
This change is motivated by issue #172. Before, the student id dialog
was populated from the *ranked_students* list, which contained all the
students that were in the list in the instant the exam was captured
(assuming the session wasn't closed and reopened, in which case ranked
students were discarded and the dialog is loaded with all the students).
Students that were added after that didn't appear in that dialog because
they didn't belong to the ranked_students list.

Now, ranked students are presented first in the dialog, and the rest
of students are presented after them. This way, newly added students
will also appear. Additionaly, only the top-5 ranked students are kept.
This improves performance.